requirements.txt
aiohttp==1.0.5
APScheduler==3.5.1
async-timeout==2.0.0
chardet==3.0.4
discord.py==0.16.12
emoji==0.5.0
multidict==3.3.2
pytz==2018.3
six==1.11.0
tzlocal==1.5.1
websockets==3.4
brawlstats
dataset